requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
pyyaml==6.0.2
PyJWT==2.9.0
redis==5.0.1
xmltodict==0.13.0